+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- VIMEET 2086 - Webinar video available available even when the happening is over
 - VIMEET-2200 - Allow more planning than participant
 - VIMEET-2155 - Password is required to change email in user account
 - VIMEET-2157 - Save import mapping